==> Trivariate_Example_Two/06_exploratory_visual_analysis.Rmd <==
---
title: "Exploratory Visual Analysis"
author: "Haaland"
date: "3/25/2020"
output: html_document
---

```{r setup, include=FALSE, message=FALSE }
knitr::opts_chunk$set(echo = TRUE)
library(knitr)
library(tidyverse)
library(MetricsWeighted)
library(MASS)
load("RData/tri_df.RData")
```

## Introduction


In this document, we are going to explore some graphical displays and consider possible statistical models and hypotheses tests. As a reminder, our data (first 15 rows) looks as follows:

```{r, message = FALSE}
kable(tri_df[1:15, ])
```

## Graphical analysis

The likely goal of the analysis is to determine what factors affect the frequency of emergency room visits (EDVisits). If we were using the raw data, we would probably do a generalized linear model with a Poisson family distribution, where each row of the data frame is for an individual subject. In that case, we could meaningfully ask questions about the mean number of ED Visits and how that is affected by roadway and particle exposure. Since we don't have access to the raw data, we are going to start from the viewpoint that we are going to investigate the distribution of ED Visits and how that is affected by the two types of exposure.


### Displaying all of the data

In the first graph, we display the frequencies on the y-axis using a log scale. We number of ED Visits is on the x-axis. Then we create facets for each combination of Roadway Exposure and Max PM2.5 Exposure. The curve in each figure is equivalent to the distribution of ED Visits within the population specified by the marginal variables. So we are interested in how the shape of the curves reveals changes in distribution based on exposure. It is interesting to compare the figures within any row or column to see the effects of one variable conditioned on a second.

```{r, fig.width = 7, fig.height = 7}
tri_df %>%
  mutate(RoadwayExposure = paste0("RoadwayExp = ", RoadwayExposure),
         MaxPMExposure = paste0("MaxPMExp = ", MaxPMExposure)) %>%
  ggplot(aes(x = EDVisits, y = log(Frequency+1))) +
    geom_point() +
    geom_line() +
    facet_grid(RoadwayExposure ~ MaxPMExposure) +
    labs(y = "Frequency") + 
    scale_x_continuous(breaks = 0:9) +
    theme(text = element_text(size = 20)) +
    scale_y_continuous(breaks = log(c(5,10,50,100,500,1000,2000)+1), labels = c(5,10,50,100,500,1000,2000))
    
```

### Comparing EdVisits distributions within a particular type of exposure

Another ways to more specifically compare effects is to put multiple curves on the same figure. An example is shown below:

```{r}
tri_df %>%
  mutate(RoadwayExposure = paste0("RoadwayExp = ", RoadwayExposure)) %>%
  ggplot(aes(x = EDVisits, y = log(Frequency+1), group = MaxPMExposure, color = MaxPMExposure)) +
    geom_point() +
    geom_line() +
    facet_wrap(~RoadwayExposure) +
    labs(y = "Frequency") + 
    scale_x_continuous(breaks = 0:9) +
    theme(text = element_text(size = 20),
          legend.position = "bottom") +
    scale_y_continuous(breaks = log(c(10,100,1000,2000)+1), labels = c(10,100,1000,2000))
    
```


### Normalization

The drawback of looking at the frequency curves is that the number of cases in each of the combinations of the two types of exposure is not related to the shape of the distributions. This is akin to the notion that in a contingency table the margins are fixed. Since we are interested in the shape of the distribution of EDVisits, we are going to normalize the frequency in each combination of the two exposure types.

We show the results of the normalization code below:

```{r}
tri_df %>%
  mutate(RoadwayExposure = paste0("RoadwayExp = ", RoadwayExposure),
         MaxPMExposure = paste0("MaxPMExp = ", MaxPMExposure)) %>%
  group_by(RoadwayExposure, MaxPMExposure) %>%
  mutate(Rate = round(Frequency / sum(Frequency),2)) %>%
  arrange(MaxPMExposure, RoadwayExposure, EDVisits)
```

Now we remake the two graphs above using the normalized data.

```{r, fig.width = 7, fig.height = 7}
tri_df %>%
  mutate(RoadwayExposure = paste0("RoadwayExp = ", RoadwayExposure),
         MaxPMExposure = paste0("MaxPMExp = ", MaxPMExposure)) %>%
  group_by(RoadwayExposure, MaxPMExposure) %>%
  mutate(Rate = Frequency / sum(Frequency)) %>%
  ggplot(aes(x = EDVisits, y = Rate)) +
    geom_point() +
    geom_line() +
    facet_grid(RoadwayExposure ~ MaxPMExposure) +
    labs(y = "Rate") + 
    scale_x_continuous(breaks = 0:9) +
    theme(text = element_text(size = 20))
    
```


```{r, fig.height = 7}
tri_df %>%
  mutate(RoadwayExposure = paste0("RoadwayExp = ", RoadwayExposure)) %>%
  group_by(RoadwayExposure, MaxPMExposure) %>%
  mutate(Rate = Frequency / sum(Frequency)) %>%
  ggplot(aes(x = EDVisits, y = Rate, group = MaxPMExposure, color = MaxPMExposure)) +
    geom_point() +
    geom_line() +
    facet_wrap(~RoadwayExposure) +
    labs(y = "Frequency") + 
    scale_x_continuous(breaks = 0:9) +
    theme(text = element_text(size = 20),
          legend.position = "bottom")
    
```

## Computing and comparing weighted means

Another way to look at the problem is that if we were considering the original data, we would be treating the number of EDVisits per individual as the response, and we would be looking for differences in means using a Poisson error distribution. We can't directly recreate that data from the three way table, but we can compute a weighted mean and standard deviation in each combination of roadway and particle exposure. Then we can compare those results graphically. We compute the weighted mean, weighted standard deviation, and weighted standard error. Then we make a simple graph. We'll do this separately for each of the two predictor features and then for the combination.

### Effect of RoadwayExposure on EDVisits

Note that the error bars include the variability due to particle exposure, which we see below is fairly large, so the error bars are pretty big. This could be fixed by getting more correct estimates from a model. I think we could still easily conclude that the effect of roadway exposure overall is fairly small.

```{r}
tri_df %>%
  group_by(RoadwayExposure) %>%
  summarize(n = sum(Frequency),
            wtd.mean = weighted_mean(EDVisits, Frequency),
            wtd.var = weighted_var(EDVisits, Frequency),
            wtd.se = sqrt(wtd.var/n),  
            error.top = min(wtd.mean + 2*wtd.se,9),
            error.bot = max(wtd.mean - 2*wtd.se,0)) %>%
  ggplot(aes(x = RoadwayExposure, y = wtd.mean)) + 
    geom_point() +
    geom_line() +
    geom_errorbar(aes(ymin = error.bot, ymax = error.top), width = .1) +
    labs(y = "Mean Number of ED Visits", 
         title = "Overall Impact of Roadway Exposure on ED Visits\n (+/- 2* standard errors)") + 
    scale_x_continuous(breaks = 0:9) +
    theme(text = element_text(size = 10)) +
    scale_y_continuous(breaks = seq(0,2, by = .25), 
                       limits = c(0,1))
```

### Effect of MaxPMExposure on EDVisits

Note that the error bars include the variability due to roadway exposure, but as we saw above there is not a lot of impact from roadway exposure, so these error bars seem pretty reasonable. The best thing, of course, would be to get estimates from a model.

```{r}
tri_df %>%
  group_by(MaxPMExposure) %>%
  summarize(n = sum(Frequency),
            wtd.mean = weighted_mean(EDVisits, Frequency),
            wtd.var = weighted_var(EDVisits, Frequency),
            wtd.se = sqrt(wtd.var/n),  
            error.top = min(wtd.mean + 2*wtd.se,9),
            error.bot = max(wtd.mean - 2*wtd.se,0)) %>%
  ggplot(aes(x = MaxPMExposure, y = wtd.mean)) + 
    geom_point() +
    geom_line() +
    geom_errorbar(aes(ymin = error.bot, ymax = error.top), width = .1) +
    labs(y = "Mean Number of ED Visits", 
         title = "Overall Impact of MaxPMExposure on ED Visits\n (+/- 2* standard errors)") + 
    scale_x_continuous(breaks = 0:9) +
    theme(text = element_text(size = 10)) +
    scale_y_continuous(breaks = seq(0,2, by = .25), 
                       limits = c(0,1))
```

### Combined Effects


```{r, fig.width = 7.5, fig.height = 7.5}
tri_df %>%
  group_by(RoadwayExposure, MaxPMExposure) %>%
  summarize(n = sum(Frequency),
            wtd.mean = weighted_mean(EDVisits, Frequency),
            wtd.var = weighted_var(EDVisits, Frequency),
            wtd.se = sqrt(wtd.var/n),  
            error.top = min(wtd.mean + 2*wtd.se,9),
            error.bot = max(wtd.mean - 2*wtd.se,0)) %>%
  group_by() %>%
  mutate(RoadwayExposure = paste0("RoadwayExp = ", RoadwayExposure)) %>%
  ggplot(aes(x = MaxPMExposure, y = wtd.mean)) + 
    geom_point() +
    geom_line() +
    geom_errorbar(aes(ymin = error.bot, ymax = error.top), width = .1) +
    facet_wrap(~RoadwayExposure) +
    labs(y = "Mean Number of ED Visits", 
         title = "Combined Impact on ED Visits\n (+/- 2* standard errors)") + 
    scale_x_continuous(breaks = 0:9) +
    theme(text = element_text(size = 20)) +
    scale_y_continuous(breaks = seq(0,2, by = .25))
```



```{r, fig.width = 7.5, fig.height = 7.5}
tri_df %>%
  group_by(RoadwayExposure, MaxPMExposure) %>%
  summarize(n = sum(Frequency),
            wtd.mean = weighted_mean(EDVisits, Frequency),
            wtd.var = weighted_var(EDVisits, Frequency),
            wtd.se = sqrt(wtd.var/n),  
            error.top = min(wtd.mean + 2*wtd.se,9),
            error.bot = max(wtd.mean - 2*wtd.se,0)) %>%
  group_by() %>%
  mutate(MaxPMExposure = paste0("MaxPMExp = ", MaxPMExposure)) %>%
  ggplot(aes(x = RoadwayExposure, y = wtd.mean)) + 
    geom_point() +
    geom_line() +
    geom_errorbar(aes(ymin = error.bot, ymax = error.top), width = .1) +
    facet_wrap(~MaxPMExposure) +
    labs(y = "Mean Number of ED Visits", 
         title = "Combined Impact on ED Visits\n (+/- 2* standard errors)") + 
    scale_x_continuous(breaks = 0:9) +
    theme(text = element_text(size = 20)) +
    scale_y_continuous(breaks = seq(0,2, by = .25))
```
